session_catalog: carve transcripts.rs (pure move)
External transcripts: cache slots + stable event ids, codex thread
projection, per-backend transcript entry parsing, activity replay,
external context-snapshot replay assembly — with tests, verbatim.
Purity-checked; tests conserved at 104.
